+++ dynsys/dynsys.py
@@ -26,10 +26,11 @@ class DynSys:
     def __init__(self,M,C,K,
                  J=None,
                  output_mtrx=None,
-                 output_names=[],
+                 output_names=None,
                  isLinear=True,
                  isSparse=False,
-                 name=None):
+                 name=None,
+                 showMsgs=True):
         """
         Dynamic systems, which may have constraints, are defined by the 
         following:
@@ -129,6 +130,9 @@ class DynSys:
         output matrices defined.
         """
         
+        if output_names is None:
+            output_names = []
+        
         self.output_names = output_names
         """
         List of string descriptions for rows of `output_mtrx`.
@@ -139,7 +143,8 @@ class DynSys:
         self._CheckSystemMatrices()
         self.CheckOutputMtrx()
         
-        print("{0} initialised.".format(self.description))
+        if showMsgs:
+            print("%s `%s` initialised." % (self.description,self.name))
         
         if isSparse:
             print("Note: sparse matrix functionality as provided by Scipy "
@@ -211,6 +216,7 @@ class DynSys:
             output_names = self.output_names
                 
         # Check list lengths agree
+        
         if len(output_names)!=output_mtrx.shape[0]:
             raise ValueError("Length of lists `output_names` "+
                              "and rows of `output_mtrx` do not agree!\n"+
@@ -255,7 +261,7 @@ class DynSys:
     def AddOutputMtrx(self,
                       output_mtrx=None,
                       output_names=None,
-                      fName=None):
+                      fName='outputs.csv'):
         """
         Appends `output_mtrx` to `outputsList`
         ***
@@ -272,6 +278,7 @@ class DynSys:
             
             output_mtrx, output_names = self.ReadOutputMtrxFromFile(fName)
             
+        # Create default output names if none provided
         if output_names is None:
             output_names = ["Response {0}".format(x) for x in range(output_mtrx.shape[0])]
             
@@ -292,6 +299,8 @@ class DynSys:
         attr_list = ["M_mtrx", "C_mtrx", "K_mtrx", "J_mtrx",
                      "M_inv", "A_inv",
                      "A_mtrx","B_mtrx"]
+        
+        print("System matrices for `{0}`:\n".format(self.name))
         
         for attr in attr_list: