+++ dynsys/tstep_results.py
@@ -304,12 +304,13 @@ class TStep_Results:
         if printProgress:
             print("Preparing results plot...")
             tic=timeit.default_timer()
+            
+        # Retrieve results from object
+        _responses = self.responses
+        _responseNames = self.responseNames
     
         # Determine total number of responses to plot
-        nResponses=0
-        for _responses in self.responses:
-            nResponses+=_responses.shape[1]
-            
+        nResponses =_responses.shape[0]
         print("nResponses to plot: {0}".format(nResponses))
         
         if nResponses == 0:
@@ -329,43 +330,42 @@ class TStep_Results:
             minVal = npy.min(self.responses)
             absmaxVal = npy.max([maxVal,minVal])
         
-        # Loop through all matrices in response list
+        # Loop through plotting all responses
         tvals = self.t
         
-        for i, _responses in enumerate(self.responses):
+        for r in range(nResponses):
             
-            _responseNames = self.responseNames[i]
+            if useCommonPlot:
+                ax = axarr
+            else:
+                ax = axarr[r]
             
-            for r in range(_responses.shape[1]):
-                
-                if useCommonPlot:
-                    ax = axarr
-                else:
-                    ax = axarr[r]
-                
-                vals = _responses[:,r]
-                print(vals.shape)
-                label_str = _responseNames[r]
-                print(label_str)
-                
-                ax.plot(tvals,vals,label=label_str)
-                                
-                ax.set_xlim([self.tstep_obj.tStart,self.tstep_obj.tEnd])
+            vals = _responses[r,:].T
+            label_str = _responseNames[r]
+            
+            ax.plot(tvals,vals,label=label_str)
+                            
+            ax.set_xlim([self.tstep_obj.tStart,self.tstep_obj.tEnd])
+            
+            if useCommonPlot:
+                if r == nResponses-1:
+                    ax.set_xlabel("Time [s]")
+            else:
                 ax.set_xlabel("Time [s]")
-                
-                if useCommonScale and not useCommonPlot:
-                    ax.set_ylim([-absmaxVal,+absmaxVal])
-                
-                if self.responseNames is not None:
-                    ax.legend(loc='best')
-                
-                # Plot horizontal lines to overlay values provided
-                # Only plot once in case of common plot
-                if (useCommonPlot and r==0) or (not useCommonPlot):
-    
-                    if y_overlay is not None:
-                        for y_val  in y_overlay:
-                            ax.axhline(y_val,color='r')
+            
+            if useCommonScale and not useCommonPlot:
+                ax.set_ylim([-absmaxVal,+absmaxVal])
+            
+            if self.responseNames is not None:
+                ax.legend(loc='right')
+            
+            # Plot horizontal lines to overlay values provided
+            # Only plot once in case of common plot
+            if (useCommonPlot and r==0) or (not useCommonPlot):
+
+                if y_overlay is not None:
+                    for y_val  in y_overlay:
+                        ax.axhline(y_val,color='r')
         
         if printProgress:
             toc=timeit.default_timer()
@@ -397,7 +397,6 @@ class TStep_Results:
                                           dofs2Plot=dofs2Plot))
         
         figs.append(self.PlotResponseResults(printProgress=printProgress,
-                                             raiseExceptions=False,
                                              useCommonPlot=useCommonPlot))
         
         return figs
@@ -413,7 +412,7 @@ class TStep_Results:
         
         """
         
-        print("Plotting PSD estimates of responses using Welch's method...")
+        print("Plotting PSD estimates of responses using periodograms...")
         
         if self.responses is None:
             raise ValueError("No response time series data avaliable!")
@@ -444,11 +443,11 @@ class TStep_Results:
         fig.legend(handles,self.responseNames,loc='upper right')
         
         # PSD plot
-        f, Pxx = scipy.signal.welch(responses,fs=fs,nperseg=nperseg)
+        f, Pxx = scipy.signal.periodogram(responses,fs)
         ax = axarr[1]
         ax.plot(f,Pxx.T)
         ax.set_xlim([0,fs/2])
-        ax.set_title("PSDs of responses")
+        ax.set_title("Periodograms of responses")
         ax.set_xlabel("Frequency (Hz)")
         
         if self.tstep_obj.name is not None:
@@ -459,6 +458,8 @@ class TStep_Results:
         fig.subplots_adjust(right=0.7)      # create space for figlegend
             
     def AnimateResults(self):
+        
+        ValueError("Unfinished - do not use!")
         
         n = self.nDOF
         
@@ -502,7 +503,8 @@ class TStep_Results:
         
     def CalcResponses(self,
                       write_results_to_file=False,
-                      results_fName="ts_results.csv"):
+                      results_fName="ts_results.csv",
+                      showMsgs=True):
         """
         Responses are obtained by pre-multiplying results by output matrices
         
@@ -523,11 +525,11 @@ class TStep_Results:
             
         # Calculate DOF statistics
         if self.calc_dof_stats:
-            self.CalcDOFStats()
+            self.CalcDOFStats(showMsgs=showMsgs)
             
         # Calculate response statistics
         if self.calc_response_stats:
-            self.CalcResponseStats()
+            self.CalcResponseStats(showMsgs=showMsgs)
             
         # Write time series results to file
         if write_results_to_file:
@@ -535,25 +537,25 @@ class TStep_Results:
             
         # Delete DOF time series data (to free-up memory)
         if not self.retainDOFTimeSeries:
-            print("Clearing DOF time series data to save memory...")
+            if showMsgs: print("Clearing DOF time series data to save memory...")
             del self.v
             del self.vdot
             del self.v2dot
         
         # Delete response time series data (to free up memory)
         if not self.retainResponseTimeSeries:
-            print("Clearing response time series data to save memory...")
+            if showMsgs: print("Clearing response time series data to save memory...")
             del self.responses
             del self.responseNames
         
-    def CalcDOFStats(self):
+    def CalcDOFStats(self,showMsgs=True):
         """
         Obtain basic statistics to describe DOF time series
         """
         
         if self.calc_dof_stats:
             
-            print("Calculating DOF statistics...")
+            if showMsgs: print("Calculating DOF statistics...")
             
             stats=[]
             
@@ -577,22 +579,21 @@ class TStep_Results:
             self.dof_stats=stats
             
         else:
-            print("calcResponseStats=False option set. Response statistics "+
-                  "will not be computed.")
+            if showMsgs:
+                print("calcResponseStats=False option set. " +
+                      "Response statistics will not be computed.")
         
         return stats
         
-    def CalcResponseStats(self):
+    def CalcResponseStats(self,showMsgs=True):
         """
         Obtain basic statistics to describe response time series
         """
         
         if self.calc_response_stats:
             
-            print("Calculating response statistics...")
-            
-            stats=[]
-            
+            if showMsgs: print("Calculating response statistics...")
+                    
             # Calculate stats for each response time series
             maxVals = npy.ravel(npy.max(self.responses,axis=1))
             minVals = npy.ravel(npy.min(self.responses,axis=1))
@@ -600,21 +601,21 @@ class TStep_Results:
             absmaxVals = npy.ravel(npy.max(npy.abs(self.responses),axis=1))
         
             # Record stats within a dict
-            d={}
-            d["max"]=maxVals
-            d["min"]=minVals
-            d["std"]=stdVals
-            d["absmax"]=absmaxVals
-            stats.append(d)
+            stats_dict={}
+            stats_dict["max"]=maxVals
+            stats_dict["min"]=minVals
+            stats_dict["std"]=stdVals
+            stats_dict["absmax"]=absmaxVals
             
             # Store within object
-            self.response_stats=stats
+            self.response_stats=stats_dict
             
         else:
-            print("calcResponseStats=False option set. Response statistics "+
-                  "will not be computed.")
+            if showMsgs:
+                print("calcResponseStats=False option set." + 
+                      "Response statistics will not be computed.")
         
-        return stats
+        return stats_dict
     
     
     def WriteResults2File(self,output_fName="timeseries_results.csv"):